+++ src/FusionStateProvider.tsx
@@ -47,7 +47,7 @@ export const useGlobalState = () => {
 /**
  * Props for the FusionStateProvider component
  */
-interface FusionStateProviderProps {
+export interface FusionStateProviderProps {
   /** Child components that will have access to fusion state */
   children: ReactNode;
   /** Initial state values to set when the provider mounts */